fix_assigned_installed_ds_tooltip_gliches
